move score component to scorecard.js, make dice total available to stateful scorecard
